Move test to examples as tt-alchemist isn't available when running tests
